Fix algorithm evaluation fixtures
Solution is to skip the container upload, this is not used during the evaluation process anyway. The other changes are just a reorganisation as this resulted in some methods not being used in the modules where they were implemented, and this resulted in some circular imports.

Closes #4963
